+++ lib/fields/checkbox.js
@@ -1,0 +1,24 @@
+'use strict';
+
+var React = require('react');
+var $ = React.DOM;
+
+
+var CheckBox = React.createClass({
+  displayName: 'CheckBox',
+
+  handleChange: function(event) {
+    var val = event.target.checked;
+    this.props.update(this.props.path, val, val);
+  },
+  render: function() {
+    return $.input({
+      name: this.props.key,
+      type: "checkbox",
+      checked: this.props.value || false,
+      onChange: this.handleChange });
+  }
+});
+
+module.exports = CheckBox;
+
