+++ lib/fields/index.js
@@ -1,0 +1,10 @@
+'use strict';
+
+module.exports = {
+    CheckBox: require('./checkbox'),
+    FileField: require('./file'),
+    InputField: require('./input'),
+    UserDefinedField: require('./user_defined'),
+    Selection: require('./selection'),
+    make: require('./make'),
+};
